xf86Crtc: Increase hardware lut size to 1024.
All the existing DDX drivers are relying on the hardcoded value of 256
for the gamma lut size. On the RK3288 the size of the lut is actually
1024.

This mismatch is causing the legacy gamma_set ioctl to return -EINVAL
making impossible to change the gamma values. As a consequence
nightshift is not working on the bitland laptop and it is also
impossible to change the color profile of the panel.

We change the hardcoded value to 1024 to mach the hw value.

Signed-off-by: Carlo Caione <carlo@endlessm.com>